requirements.txt
streamlit>=1.32.0
torch>=2.1.0
transformers>=4.38.0
pillow>=10.0.0
accelerate>=0.27.0